+++ pybamm/expression_tree/interpolant.py
@@ -42,6 +42,10 @@ class Interpolant(pybamm.Function):
         extrapolate=True,
         entries_string=None,
     ):
+
+        # set default dimension value
+        self.dimension = 1
+
         if isinstance(x, (tuple, list)) and len(x) == 2:
             interpolator = interpolator or "linear"
             if interpolator != "linear":
@@ -59,6 +63,7 @@ class Interpolant(pybamm.Function):
                 x1 = x
                 x = [x]
             x2 = None
+
         if x1.shape[0] != y.shape[0]:
             raise ValueError(
                 "len(x1) should equal y=shape[0], "
@@ -84,6 +89,7 @@ class Interpolant(pybamm.Function):
 
         if interpolator == "linear":
             if len(x) == 1:
+                self.dimension = 1
                 if extrapolate is False:
                     interpolating_function = interpolate.interp1d(
                         x1, y.T, bounds_error=False, fill_value=np.nan
@@ -93,7 +99,10 @@ class Interpolant(pybamm.Function):
                         x1, y.T, bounds_error=False, fill_value="extrapolate"
                     )
             elif len(x) == 2:
+                self.dimension = 2
                 interpolating_function = interpolate.interp2d(x1, x2, y)
+            else:
+                raise ValueError("Invalid dimension of x: {0}".format(len(x)))
         elif interpolator == "pchip":
             interpolating_function = interpolate.PchipInterpolator(
                 x1, y, extrapolate=extrapolate
@@ -161,5 +170,16 @@ class Interpolant(pybamm.Function):
                 children_eval_flat.append(child.flatten())
             else:
                 children_eval_flat.append(child)
+        if self.dimension == 1:
+            return self.function(*children_eval_flat).flatten()[:, np.newaxis]
+        elif self.dimension == 2:
+            res = self.function(*children_eval_flat)
+            if res.ndim > 1:
+                return np.diagonal(res)[:, np.newaxis]
+            else:
+                # raise ValueError("Invalid children dimension: {0}".format(res.ndim))
+                return res[:, np.newaxis]
+        else:  # pragma: no cover
+            raise ValueError("Invalid dimension: {0}".format(self.dimension))
 
-        return self.function(*children_eval_flat).flatten()[:, np.newaxis]
+