+++ pybamm/models/submodels/interface/kinetics/base_kinetics.py
@@ -1,7 +1,6 @@
 #
 # Base kinetics class
 #
-
 import pybamm
 from ..base_interface import BaseInterface