+++ tests/unit/test_expression_tree/test_symbol.py
@@ -66,16 +66,31 @@ class TestSymbol(unittest.TestCase):
         a = pybamm.Symbol(
             "a",
             domain="test",
-            auxiliary_domains={"secondary": "secondary", "tertiary": "tertiary"},
+            auxiliary_domains={
+                "secondary": "sec",
+                "tertiary": "tert",
+                "quaternary": "quat",
+            },
         )
         self.assertEqual(a.domain, ["test"])
         self.assertEqual(
-            a.auxiliary_domains, {"secondary": ["secondary"], "tertiary": ["tertiary"]}
+            a.auxiliary_domains,
+            {"secondary": ["sec"], "tertiary": ["tert"], "quaternary": ["quat"]},
         )
+        self.assertEqual(a.secondary_domain, ["sec"])
+        self.assertEqual(a.tertiary_domain, ["tert"])
+        self.assertEqual(a.tertiary_domain, ["tert"])
+        self.assertEqual(a.quaternary_domain, ["quat"])
         self.assertEqual(
             a.domains,
-            {"primary": ["test"], "secondary": ["secondary"], "tertiary": ["tertiary"]},
+            {
+                "primary": ["test"],
+                "secondary": ["sec"],
+                "tertiary": ["tert"],
+                "quaternary": ["quat"],
+            },
         )
+
         a = pybamm.Symbol("a", domain=["t", "e", "s"])
         self.assertEqual(a.domain, ["t", "e", "s"])
         with self.assertRaises(TypeError):