+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_spm.py
@@ -15,6 +15,14 @@ class TestSPM(BaseUnitTestLithiumIon, unittest.TestCase):
         with self.assertRaisesRegex(pybamm.OptionError, "electrolyte conductivity"):
             pybamm.lithium_ion.SPM(options)
 
+    def test_kinetics_options(self):
+        options = {
+            "surface form": "false",
+            "intercalation kinetics": "Marcus-Hush-Chidsey",
+        }
+        with self.assertRaisesRegex(pybamm.OptionError, "Inverse kinetics"):
+            pybamm.lithium_ion.SPM(options)
+
     def test_new_model(self):
         model = pybamm.lithium_ion.SPM({"thermal": "x-full"})
         new_model = model.new_copy()